feat: split methods to single functions
Not only are the functions now splited into separate functions, the build process is aligned as
well. The documentation got an overall fix, jsdoc compatibility is better and a lot of
configurations are now included inside the package.json file.
