+++ .jsdoc.json
@@ -1,0 +1,11 @@
+{
+    "source": {
+        "include": ["src", "README.md"],
+        "includePattern": ".js$",
+        "excludePattern": "(node_modules/|.test.js)"
+    },
+    "opts": {
+        "destination": "docs",
+        "recurse": true
+    }
+}
