+++ src/aleaRNGFactory.d.ts
@@ -1,0 +1,3 @@
+/*! The MIT License (MIT) - Copyright (c) 2016-2018 Martin Helmut Fieber */
+import { aleaRNGFactory } from './index';
+export default aleaRNGFactory;
