+++ src/murmurhash2_x86_32.d.ts
@@ -1,0 +1,3 @@
+/*! The MIT License (MIT) - Copyright (c) 2016-2018 Martin Helmut Fieber */
+import { murmurhash2_x86_32 } from './index';
+export default murmurhash2_x86_32;
